requirements.txt
# Core Streamlit and PDF processing
streamlit>=1.28.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0

# AI/ML Libraries
openai>=1.0.0
google-generativeai>=0.3.0
huggingface-hub>=0.17.0
transformers>=4.35.0

# Vector processing and embeddings
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.21.0

# RAG and document processing
langchain>=0.0.300
faiss-cpu>=1.7.4

# HTTP requests
requests>=2.31.0
httpx>=0.25.0

# Utilities
json5>=0.9.0